refactor: put useInput to hooks folder
